requirements.txt
asgiref==3.7.2
boto3==1.28.8
botocore==1.31.8
colorama==0.4.6
Django==4.2.3
django-ses==3.5.0
djangorestframework==3.14.0
jmespath==1.0.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.16
